GOCasting: Split includes for spells from Unit.h to own file and add it to GameObject.h
